requirements.txt
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coverage==7.4.3
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
mobizon-client==0.1.2
msgpack==1.0.8
oauthlib==3.2.2
packaging==23.2
phonenumberslite==8.13.30
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
python3-openid==3.2.0
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.2
